refactor: rename navbar component
rename navbar component from NavBar.vue to MainNavbar.vue
